Extract wallet table component
Extract wallet table component in order to reuse the component used to select the execution multisig also when selectinf executive committee on a gasless dao
