+++ src/components/multisigWallets/index.tsx
@@ -1,17 +1,20 @@
 import React, {useEffect, useRef} from 'react';
 import {useFieldArray, useFormContext, useWatch} from 'react-hook-form';
-import {Dropdown, Label, ListItemAction} from '@aragon/ods-old';
-import {Button, IconType, AlertInline} from '@aragon/ods';
+import {Label} from '@aragon/ods-old';
+import {AlertInline} from '@aragon/ods';
 import {useTranslation} from 'react-i18next';
 import styled from 'styled-components';
-import {Address, useEnsName} from 'wagmi';
 
 import {useAlertContext} from 'context/alert';
-import {useNetwork} from 'context/network';
-import useScreen from 'hooks/useScreen';
 import {useWallet} from 'hooks/useWallet';
-import {CHAIN_METADATA} from 'utils/constants';
-import {Row} from './row';
+import {MultisigWalletField} from './row';
+import {WalletTable} from './walletTable';
+import {Web3Address} from 'utils/library';
+import {validateWeb3Address} from 'utils/validators';
+
+import {useNetwork} from '../../context/network';
+import {Address, useEnsName} from 'wagmi';
+import {CHAIN_METADATA} from '../../utils/constants';
 
 export const MultisigWallets = () => {
   const {t} = useTranslation();
@@ -28,7 +31,8 @@ export const MultisigWallets = () => {
 
   const {control, trigger, setFocus} = useFormContext();
   const multisigWallets = useWatch({name: 'multisigWallets', control});
-  const {fields, update, replace, append, remove} = useFieldArray({
+
+  const {fields, replace, append, remove} = useFieldArray({
     control,
     name: 'multisigWallets',
   });
@@ -81,26 +85,27 @@ export const MultisigWallets = () => {
     }, 50);
   };
 
-  // reset wallet
-  const handleResetEntry = (index: number) => {
-    update(index, {address: '', ensName: ''});
-    alert(t('alert.chip.resetAddress'));
-    trigger('multisigWallets');
-  };
+  const addressValidator = async (wallet: Web3Address, index: number) => {
+    let validationResult = await validateWeb3Address(
+      wallet,
+      t('errors.required.walletAddress'),
+      t
+    );
 
-  // reset all wallets
-  const handleResetAll = () => {
-    controlledWallets.forEach((_, index) => {
-      // skip the first one because is the own address
-      if (index > 0) {
-        update(index, {address: '', ensName: ''});
-      }
-    });
-    alert(t('alert.chip.resetAllAddresses'));
-    trigger('multisigWallets');
+    if (multisigWallets) {
+      multisigWallets.forEach(
+        ({address, ensName}: MultisigWalletField, itemIndex: number) => {
+          if (
+            (address === wallet.address || ensName === wallet.ensName) &&
+            itemIndex !== index
+          ) {
+            validationResult = t('errors.duplicateAddress');
+          }
+        }
+      );
+    }
+    return validationResult;
   };
-
-  const {isMobile} = useScreen();
 
   return (
     <Container>
@@ -111,80 +116,14 @@ export const MultisigWallets = () => {
           renderHtml
         />
       </DescriptionContainer>
-      <TableContainer>
-        {!isMobile && (
-          <TableTitleContainer>
-            <Title>{t('labels.whitelistWallets.address')}</Title>
-          </TableTitleContainer>
-        )}
-        {controlledWallets.map((field, index) => (
-          <div key={field.id}>
-            {(!isMobile || (isMobile && index !== 0)) && <Divider />}
-            <Row
-              index={index}
-              onResetEntry={handleResetEntry}
-              onDeleteEntry={handleDeleteEntry}
-            />
-          </div>
-        ))}
-        <Divider />
-        <ActionsContainer>
-          <TextButtonsContainer>
-            <Button variant="tertiary" size="lg" onClick={handleAdd}>
-              {t('labels.whitelistWallets.addAddress')}
-            </Button>
-            {/*
-          To be enabled when csv functionality is there
-          <ButtonText
-            label={t('labels.whitelistWallets.uploadCSV')}
-            mode="ghost"
-            size="large"
-            onClick={() => alert('upload CSV here')}
-          /> */}
-          </TextButtonsContainer>
-          <Dropdown
-            side="bottom"
-            align="start"
-            sideOffset={4}
-            trigger={
-              <Button
-                size="lg"
-                variant="tertiary"
-                iconLeft={IconType.DOTS_VERTICAL}
-                data-testid="trigger"
-              />
-            }
-            listItems={[
-              {
-                component: (
-                  <ListItemAction
-                    title={t('labels.whitelistWallets.resetAllEntries')}
-                    bgWhite
-                  />
-                ),
-                callback: handleResetAll,
-              },
-              {
-                component: (
-                  <ListItemAction
-                    title={t('labels.whitelistWallets.deleteAllEntries')}
-                    bgWhite
-                  />
-                ),
-                callback: handleDeleteAll,
-              },
-            ]}
-          />
-        </ActionsContainer>
-        <Divider />
-        <SummaryContainer>
-          <Title>{t('labels.summary')}</Title>
-          <TotalWalletsContainer>
-            <Text>{t('labels.whitelistWallets.totalWallets')}</Text>
-            <Title>{controlledWallets.length}</Title>
-          </TotalWalletsContainer>
-        </SummaryContainer>
-      </TableContainer>
+      <WalletTable
+        controlledWallets={controlledWallets}
+        handleAdd={handleAdd}
+        handleDeleteEntry={handleDeleteEntry}
+        handleDeleteAll={handleDeleteAll}
+        rowValidator={addressValidator}
+        walletFieldName={'multisigWallets'}
+      />
       <AlertInline
         message={t('createDAO.step3.multisigMembersWalletAlert')}
         variant="info"
@@ -199,31 +138,31 @@ const Container = styled.div.attrs(() => ({
 const DescriptionContainer = styled.div.attrs(() => ({
   className: 'space-y-1 flex flex-col',
 }))``;
-const TableContainer = styled.div.attrs(() => ({
+styled.div.attrs(() => ({
   className: 'rounded-xl bg-neutral-0 flex flex-col',
 }))``;
-const TableTitleContainer = styled.div.attrs(() => ({
+styled.div.attrs(() => ({
   className: 'mx-6 mt-6 mb-3',
 }))``;
-const Title = styled.p.attrs({
+styled.p.attrs({
   className: 'ft-text-base xl:font-semibold font-semibold text-neutral-800',
 })``;
-const Text = styled.p.attrs({
+styled.p.attrs({
   className: 'ft-text-base  text-neutral-600',
 })``;
-const Divider = styled.div.attrs(() => ({
+styled.div.attrs(() => ({
   className: 'flex bg-neutral-50 h-0.5',
 }))``;
-const ActionsContainer = styled.div.attrs(() => ({
+styled.div.attrs(() => ({
   className: 'flex xl:px-6 xl:py-3 p-4 place-content-between',
 }))``;
-const TextButtonsContainer = styled.div.attrs(() => ({
+styled.div.attrs(() => ({
   className: 'flex gap-4',
 }))``;
 
-const SummaryContainer = styled.div.attrs(() => ({
+styled.div.attrs(() => ({
   className: 'flex xl:p-6 p-4 flex-col space-y-3',
 }))``;
-const TotalWalletsContainer = styled.div.attrs(() => ({
+styled.div.attrs(() => ({
   className: 'flex place-content-between',
 }))``;
